Modify example app to fix placeholders
When the user entered some text in any of the cells and left the cell empty after that, the placeholder was not appearing any more. With these changes, the placeholder appears again when the text of the cell is empty.

Fixes the following issue: #162 placeholder text disappear when textfield is erased